pyproject.toml
[tool.poetry]
name = "remi"
version = "0.1.0"
description = "A personal Discord bot for my Discord server"
authors = ["PythonTryHard <31789326+PythonTryHard@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
hikari = {extras = ["speedups", "server"], version = "^2.0.0.dev108"}
hikari-lightbulb = "^2.2.1"
uvloop = [{ version = "^0.16.0", platform = "linux" }]
python-dotenv = "^0.20.0"
tzlocal = "^4.2"
loguru = "^0.6.0"
SQLAlchemy = "^1.4.34"
aiosqlite = "^0.17.0"
rich = "^12.2.0"
click = "^8.1.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.16.0"
black = "^22.3.0"
pudb = "^2022.1.1"
pylint = "^2.13.4"
pytest = "^7.1.1"
pytest-repeat = "^0.9.1"
isort = "^5.10.1"
memory-profiler = "^0.60.0"
matplotlib = "^3.5.1"
cookiecutter = "^1.7.3"

[tool.poetry.scripts]
remi = 'remi:__main__.main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = [
    "missing-function-docstring",
    "missing-module-docstring",
    "missing-class-docstring",
    "unused-import",
    "unused-argument",
    "expression-not-assigned",
    "too-few-public-methods"
]
